FFC-52 [fix] Added 2 signs after period, added level to json output

+++ app/domain/bigdata/GroupingSets.java
@@ -16,6 +16,6 @@ public class GroupingSets implements Serializable {
 
     long statIndicatorId;
     long groupingSetId;
-    long level;
+    int level;
 
 }
